Use Module to Set Breakpoint

Module of Shiny app needed to enable functionality to setting breakpoint. Must be used in the server part of Shiny app or in the function which is then used in the server part of app.

Usage

shinybreakpointServer(
  keyEvent = "F4",
  id = "shinybreakpoint",
  varName = "....envirr"
)

Arguments

keyEvent

key to run modal dialog with the functionality to set breakpoint. "F4" by default.

id

namespace used for all inputs in this module. "shinybreakpoint" by default. Change if in the app some other module is used which already has "shinybreakpoint" namespace.

varName

setting breakpoint is not equal to inserting only browser(), but also additional code, one of which is assignment operation. This parameter determines the variable name to assign value. "....envirr" by default. Change if this name is already in use somewhere in the app.

Value

Used for side effect - adds modal dialog to the Shiny app with the options to set breakpoint. Modal dialog is shown when the key specified in keyEvent is pressed.

Details

One of the core concepts which founds this module is the necessity to re-run the objects present in the server part of app. This is possible only when these objects do not live directly in the server, but in the function which is then used in server. This naturally harmonizes with the modules, but it needs the separate function for objects which would be used directly in the server (see example).

Examples

# To run example, copy-paste to file, save the file
# and run the app. Then press "F4" to open the modal dialog.
# Make sure 'bslib' package is installed.

if (FALSE) {

# install.packages("bslib") if needed
library(shiny)

appServer <- function(input, output, session) {
  observe({
    input$num
  })
}

shinyApp(
  ui = fluidPage(
    theme = bslib::bs_theme(),
    numericInput("num", "Num", 0)
  ),
  server = function(input, output, session) {
    shinybreakpoint::shinybreakpointServer()
    appServer(input, output, session)
  }
)
}